requirements.txt
fastapi
uvicorn
google-generativeai
googlemaps
requests
python-dotenv
pytest
httpx
